very basic init added. Need to still add parameters.
